+++ my_drivers/Src/my_uart.c
@@ -23,11 +23,11 @@ void my_uart1_init ( void )
 	GPIOA->OTYPER	|= GPIO_OTYPER_OT3 ; 		// Choose Open_drain for RX to avoid P-MOS transistors problems and assuming push-pull expected on device
 }
 
-void my_uart1_on ( uint16_t my_sysclock , uint16_t br )
+void my_uart1_on ( uint32_t my_sysclock , uint32_t br )
 {
 	RCC->IOPENR 	|= RCC_IOPENR_GPIOAEN ;		// Activate clock in GPIOA
 	RCC->APBENR2	|= RCC_APBENR2_USART1EN ;
-	USART1->BRR 	=  my_sysclock / br ;			// Activate clock for UART
+	USART1->BRR 	=  (uint16_t) ( my_sysclock / br ) ;			// Activate clock for UART
 	USART1->CR1 	|=  USART_CR1_UE ;
 	USART1->CR1 	|=  USART_CR1_TE ;
 	USART1->CR1 	|=  USART_CR1_RE ;
@@ -43,11 +43,11 @@ void my_uart2_init ( void )
 	GPIOA->OTYPER	|= GPIO_OTYPER_OT3 ; 		// Choose Open_drain for RX to avoid P-MOS transistors problems and assuming push-pull expected on device
 }
 
-void my_uart2_on ( uint16_t my_sysclock , uint16_t br )
+void my_uart2_on ( uint32_t my_sysclock , uint32_t br )
 {
 	RCC->IOPENR 	|= RCC_IOPENR_GPIOAEN ;		// Activate clock in GPIOA
 	RCC->APBENR1	|= RCC_APBENR1_USART2EN ;
-	USART2->BRR 	= my_sysclock / br ;			// Activate clock for UART
+	USART2->BRR 	= (uint16_t) ( my_sysclock / br ) ;			// Activate clock for UART
 	USART2->CR1 	|=  USART_CR1_UE ;
 	USART2->CR1 	|=  USART_CR1_TE ;
 	USART2->CR1 	|=  USART_CR1_RE ;
